Allow build DALI for different architectures as well (#1397)
- allows the specification of DALI architecture as a build parameter,
  currently, only x86_64 is supported
- allows passing CUDA_TARGET_ARCHS as a build parameter
- adds an option to skip building of prebuild DALI TF plugin

+++ docker/build.sh
@@ -10,12 +10,15 @@ NVIDIA_BUILD_ID=[default 12345]
 CREATE_WHL=[default YES]
 CREATE_RUNNER=[default NO]
 BUILD_TF_PLUGIN=[default NO]
+PREBUILD_TF_PLUGINS=[default YES]
 DALI_BUILD_FLAVOR=[default is empty]
 CMAKE_BUILD_TYPE=[default is Release]
 BUILD_INHOST=[create build dir with object outside docker, just mount it as a volume, default is YES]
 REBUILD_BUILDERS=[default is NO]
 REBUILD_MANYLINUX=[default is NO]
 DALI_BUILD_DIR=[default is build-docker-\${CMAKE_BUILD_TYPE}-\${PYV}-\${CUDA_VERSION}]
+ARCH=[default is x86_64]
+WHL_PLATFORM_NAME=[default is manylinux1_x86_64]
 
 where:
     -h  show this help text"
@@ -34,6 +37,7 @@ done
 shift $((OPTIND - 1))
 
 #########Set Me###############
+export ARCH=${ARCH:-x86_64}
 export PYVER=${PYVER:-3.6}
 export PYV=${PYVER/./}
 export CUDA_VERSION=${CUDA_VERSION:-10}
@@ -61,18 +65,21 @@ export BUILD_INHOST=${BUILD_INHOST:-YES}
 export REBUILD_BUILDERS=${REBUILD_BUILDERS:-NO}
 export REBUILD_MANYLINUX=${REBUILD_MANYLINUX:-NO}
 export BUILD_TF_PLUGIN=${BUILD_TF_PLUGIN:-NO}
-export DALI_BUILD_DIR=${DALI_BUILD_DIR:-build-docker-${CMAKE_BUILD_TYPE}-${PYV}-${CUDA_VERSION}}
+export PREBUILD_TF_PLUGINS=${PREBUILD_TF_PLUGINS:-YES}
+export DALI_BUILD_DIR=${DALI_BUILD_DIR:-build-docker-${CMAKE_BUILD_TYPE}-${PYV}-${CUDA_VERSION}}_${ARCH}
+export WHL_PLATFORM_NAME=${WHL_PLATFORM_NAME:-manylinux1_x86_64}
 #################################
-export DEPS_IMAGE=nvidia/dali:cu${CUDA_VERSION}.deps
-export CUDA_DEPS_IMAGE=nvidia/dali:cuda${CUDA_VERSION}.toolkit
-export BUILDER=nvidia/dali:py${PYV}_cu${CUDA_VERSION}.build
-export BUILDER_WHL=nvidia/dali:py${PYV}_cu${CUDA_VERSION}.build_whl
+export BASE_NAME=manylinux3_${ARCH}
+export DEPS_IMAGE=nvidia/dali:cu${CUDA_VERSION}_${ARCH}.deps
+export CUDA_DEPS_IMAGE=nvidia/dali:cuda${CUDA_VERSION}_${ARCH}.toolkit
+export BUILDER=nvidia/dali:py${PYV}_cu${CUDA_VERSION}_${ARCH}.build
+export BUILDER_WHL=nvidia/dali:py${PYV}_cu${CUDA_VERSION}_${ARCH}.build_whl
 export BUILDER_DALI_TF_BASE_MANYLINUX1=nvidia/dali:py${PYV}_cu${CUDA_VERSION}.build_tf_base_manylinux1
 export BUILDER_DALI_TF_BASE_MANYLINUX2010=nvidia/dali:py${PYV}_cu${CUDA_VERSION}.build_tf_base_manylinux2010
 export BUILDER_DALI_TF_BASE_WITH_WHEEL=nvidia/dali:py${PYV}_cu${CUDA_VERSION}.build_tf_base_with_whl
 export BUILDER_DALI_TF_MANYLINUX1=nvidia/dali:py${PYV}_cu${CUDA_VERSION}.build_tf_manylinux1
 export BUILDER_DALI_TF_MANYLINUX2010=nvidia/dali:py${PYV}_cu${CUDA_VERSION}.build_tf_manylinux2010
-export BUILDER_DALI_TF_SDIST=nvidia/dali:py${PYV}_cu${CUDA_VERSION}.build_tf_sdist
+export BUILDER_DALI_TF_SDIST=nvidia/dali:py${PYV}_cu${CUDA_VERSION}_${ARCH}.build_tf_sdist
 export RUN_IMG=nvidia/dali:py${PYV}_cu${CUDA_VERSION}.run
 export GIT_SHA=$(git rev-parse HEAD)
 export DALI_TIMESTAMP=$(date +%Y%m%d)
@@ -80,31 +87,33 @@ export DALI_TIMESTAMP=$(date +%Y%m%d)
 set -o errexit
 
 # build manylinux3 if needed
-if [[ "$(docker images -q manylinux3_x86_64 2> /dev/null)" == "" || "$REBUILD_MANYLINUX" != "NO" ]]; then
+if [[ "$REBUILD_MANYLINUX" != "NO" || "$(docker images -q ${BASE_NAME} 2> /dev/null)" == "" ]]; then
     pushd ../third_party/manylinux/
     git checkout 96b47a25673b33c728e49099a3a6b1bf503a18c2 || echo -e "Did you forget to \`git clone --recursive\`? Try this:\n" \
                                                                     "  git submodule sync --recursive && \n" \
                                                                     "  git submodule update --init --recursive && \n"
-    git am ../../docker/0001-An-approximate-manylinux3.patch
+    git am ../../docker/0001-An-approximate-manylinux3_${ARCH}.patch
     PLATFORM=$(uname -m) TRAVIS_COMMIT=latest ./build.sh
     popd
 fi
 
 pushd ../
 # build deps image if needed
-if [[ "$(docker images -q ${DEPS_IMAGE} 2> /dev/null)" == "" || "$(docker images -q ${CUDA_DEPS_IMAGE} 2> /dev/null)" == "" || "$REBUILD_BUILDERS" != "NO" ]]; then
+if [[ "$REBUILD_BUILDERS" != "NO" || "$(docker images -q ${DEPS_IMAGE} 2> /dev/null)" == "" || "$(docker images -q ${CUDA_DEPS_IMAGE} 2> /dev/null)" == "" ]]; then
     echo "Build deps: " ${DEPS_IMAGE}
     docker build -t ${CUDA_DEPS_IMAGE} -f docker/Dockerfile.cuda${CUDA_VERSION}.deps .
-    docker build -t ${DEPS_IMAGE} --build-arg "FROM_IMAGE_NAME"=manylinux3_x86_64 --build-arg "CUDA_IMAGE=${CUDA_DEPS_IMAGE}" -f docker/Dockerfile.deps .
+    docker build -t ${DEPS_IMAGE} --build-arg "FROM_IMAGE_NAME"=${BASE_NAME} --build-arg "CUDA_IMAGE=${CUDA_DEPS_IMAGE}" -f docker/Dockerfile.deps .
 fi
 
 # build builder image if needed
-if [[ "$(docker images -q ${BUILDER} 2> /dev/null)" == "" || "$(docker images -q ${BUILDER_DALI_TF_BASE_MANYLINUX1} 2> /dev/null)" == "" || "$(docker images -q ${BUILDER_DALI_TF_BASE_MANYLINUX2010} 2> /dev/null)" || "$REBUILD_BUILDERS" != "NO" ]]; then
+if [[ "$REBUILD_BUILDERS" != "NO" || "$(docker images -q ${BUILDER} 2> /dev/null)" == "" ]]; then
     echo "Build light image:" ${BUILDER}
     docker build -t ${BUILDER} --build-arg "DEPS_IMAGE_NAME=${DEPS_IMAGE}" --build-arg "PYVER=${PYVER}" --build-arg "PYV=${PYV}" --build-arg "NVIDIA_BUILD_ID=${NVIDIA_BUILD_ID}" \
                                --build-arg "NVIDIA_DALI_BUILD_FLAVOR=${DALI_BUILD_FLAVOR}" --build-arg "GIT_SHA=${GIT_SHA}" --build-arg "DALI_TIMESTAMP=${DALI_TIMESTAMP}" \
                                --build-arg "CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE}" --target builder -f docker/Dockerfile .
+fi
 
+if [[ "$BUILD_TF_PLUGIN" == "YES" && "${PREBUILD_TF_PLUGINS}" == "YES" && ("$REBUILD_BUILDERS" != "NO" || "$(docker images -q ${BUILDER_DALI_TF_BASE_MANYLINUX1} 2> /dev/null)" == "" || "$(docker images -q ${BUILDER_DALI_TF_BASE_MANYLINUX2010} 2> /dev/null)"  == "") ]]; then
     echo "Build DALI TF base (manylinux1)"
     TF_CUSTOM_OP_IMAGE_MANYLINUX1="tensorflow/tensorflow:custom-op-gpu-ubuntu14"
     docker build -t ${BUILDER_DALI_TF_BASE_MANYLINUX1} \
@@ -122,15 +131,18 @@ if [[ "$(docker images -q ${BUILDER} 2> /dev/null)" == "" || "$(docker images -q
 
 fi
 
-if [ "$BUILD_INHOST" = "YES" ]; then
+if [ "$BUILD_INHOST" == "YES" ]; then
     # build inside the source tree
-    docker run --rm -u 1000:1000 -v $(pwd):/opt/dali ${BUILDER} /bin/bash -c "mkdir -p /opt/dali/${DALI_BUILD_DIR} && \
+    docker run --rm -u $(id -u ${USER}):$(id -g ${USER}) -v $(pwd):/opt/dali ${BUILDER} /bin/bash -c "mkdir -p /opt/dali/${DALI_BUILD_DIR} && \
                                         cd /opt/dali/${DALI_BUILD_DIR} &&         \
                                         rm -rf /opt/dali/${DALI_BUILD_DIR}/nvidia* && \
+                                        ARCH=${ARCH}                              \
+                                        WHL_PLATFORM_NAME=${WHL_PLATFORM_NAME}    \
+                                        CUDA_TARGET_ARCHS=\"${CUDA_TARGET_ARCHS}\"\
                                         CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE}      \
                                         BUILD_TEST=${BUILD_TEST}                  \
                                         BUILD_BENCHMARK=${BUILD_BENCHMARK}        \
-                                        BUILD_NVTX=${BUILD_NVTX}        \
+                                        BUILD_NVTX=${BUILD_NVTX}                  \
                                         BUILD_PYTHON=${BUILD_PYTHON}              \
                                         BUILD_LMDB=${BUILD_LMDB}                  \
                                         BUILD_JPEG_TURBO=${BUILD_JPEG_TURBO}      \
@@ -151,13 +163,37 @@ if [ "$BUILD_INHOST" = "YES" ]; then
                                         cp /wheelhouse/* ./"
 else
     echo "Build image:" ${BUILDER_WHL}
-    docker build -t ${BUILDER_WHL} --build-arg "DEPS_IMAGE_NAME=${DEPS_IMAGE}" --build-arg "PYVER=${PYVER}" --build-arg "PYV=${PYV}" --build-arg "NVIDIA_BUILD_ID=${NVIDIA_BUILD_ID}" \
-                                   --build-arg "NVIDIA_DALI_BUILD_FLAVOR=${DALI_BUILD_FLAVOR}" --build-arg "GIT_SHA=${GIT_SHA}" --build-arg "DALI_TIMESTAMP=${DALI_TIMESTAMP}" \
-                                   --build-arg "CMAKE_BUILD_TYPE=${BUILD_TYPE}" --cache-from "${BUILDER}" .
+    docker build -t ${BUILDER_WHL} --build-arg "DEPS_IMAGE_NAME=${DEPS_IMAGE}"             \
+                                   --build-arg "PYVER=${PYVER}"                            \
+                                   --build-arg "PYV=${PYV}"                                \
+                                   --build-arg "ARCH=${ARCH}"                              \
+                                   --build-arg "WHL_PLATFORM_NAME=${WHL_PLATFORM_NAME}"    \
+                                   --build-arg "CUDA_TARGET_ARCHS=${CUDA_TARGET_ARCHS}"    \
+                                   --build-arg "CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE}"      \
+                                   --build-arg "BUILD_TEST=${BUILD_TEST}"                  \
+                                   --build-arg "BUILD_BENCHMARK=${BUILD_BENCHMARK}"        \
+                                   --build-arg "BUILD_NVTX=${BUILD_NVTX}"                  \
+                                   --build-arg "BUILD_PYTHON=${BUILD_PYTHON}"              \
+                                   --build-arg "BUILD_LMDB=${BUILD_LMDB}"                  \
+                                   --build-arg "BUILD_JPEG_TURBO=${BUILD_JPEG_TURBO}"      \
+                                   --build-arg "BUILD_NVJPEG=${BUILD_NVJPEG}"              \
+                                   --build-arg "BUILD_LIBTIFF=${BUILD_LIBTIFF}"            \
+                                   --build-arg "BUILD_NVOF=${BUILD_NVOF}"                  \
+                                   --build-arg "BUILD_NVDEC=${BUILD_NVDEC}"                \
+                                   --build-arg "BUILD_NVML=${BUILD_NVML}"                  \
+                                   --build-arg "VERBOSE_LOGS=${VERBOSE_LOGS}"              \
+                                   --build-arg "WERROR=${WERROR}"                          \
+                                   --build-arg "BUILD_WITH_ASAN=${BUILD_WITH_ASAN}"        \
+                                   --build-arg "NVIDIA_BUILD_ID=${NVIDIA_BUILD_ID}"        \
+                                   --build-arg "GIT_SHA=${GIT_SHA}"                        \
+                                   --build-arg "DALI_TIMESTAMP=${DALI_TIMESTAMP}"          \
+                                   --build-arg "NVIDIA_DALI_BUILD_FLAVOR=${DALI_BUILD_FLAVOR}" \
+                                   --cache-from "${BUILDER}"                               \
+                                   -f docker/Dockerfile .
 fi
 
 
-if [ "$CREATE_RUNNER" = "YES" ]; then
+if [ "$CREATE_RUNNER" == "YES" ]; then
     echo "Runner image:" ${RUN_IMG}
     echo "You can run this image with DALI installed inside, keep in mind to install neccessary FW package as well"
     if [ ${CUDA_VERSION} == "9" ] ; then
@@ -190,8 +226,8 @@ fi
 
 mkdir -p ./wheelhouse/
 
-if [ "$CREATE_WHL" = "YES" ]; then
-    if [ "$BUILD_INHOST" = "YES" ]; then
+if [ "$CREATE_WHL" == "YES" ]; then
+    if [ "$BUILD_INHOST" == "YES" ]; then
         cp $(pwd)/${DALI_BUILD_DIR}/nvidia* ./wheelhouse/
     else
         export CONTAINER="extract-tmp"
@@ -202,7 +238,7 @@ if [ "$CREATE_WHL" = "YES" ]; then
 fi
 
 
-if [ "$BUILD_TF_PLUGIN" = "YES" ]; then
+if [[ "$CREATE_WHL" == "YES" && "$BUILD_TF_PLUGIN" = "YES" ]]; then
 
     mkdir -p dali_tf_plugin/whl
     cp ./wheelhouse/*.whl dali_tf_plugin/whl/
@@ -226,31 +262,33 @@ if [ "$BUILD_TF_PLUGIN" = "YES" ]; then
 #             /bin/bash build_in_custom_op_docker.sh"
     # else
 
-    echo "Build image:" ${BUILDER_DALI_TF_MANYLINUX1}
-    docker build -t ${BUILDER_DALI_TF_MANYLINUX1} -f docker/Dockerfile_dali_tf \
-        --build-arg "TF_CUSTOM_OP_BUILDER_IMAGE=${BUILDER_DALI_TF_BASE_MANYLINUX1}" \
-        .
-    export DALI_TF_BUILDER_CONTAINER_MANYLINUX1="extract_dali_tf_prebuilt_manylinux1"
-    nvidia-docker run --name ${DALI_TF_BUILDER_CONTAINER_MANYLINUX1} ${BUILDER_DALI_TF_MANYLINUX1} /bin/bash -c 'source /opt/dali/dali_tf_plugin/build_in_custom_op_docker.sh'
-    docker cp "${DALI_TF_BUILDER_CONTAINER_MANYLINUX1}:/prebuilt/." "prebuilt_manylinux1"
-    docker rm -f "${DALI_TF_BUILDER_CONTAINER_MANYLINUX1}"
-
-    echo "Build image:" ${BUILDER_DALI_TF_MANYLINUX2010}
-    docker build -t ${BUILDER_DALI_TF_MANYLINUX2010} -f docker/Dockerfile_dali_tf \
-        --build-arg "TF_CUSTOM_OP_BUILDER_IMAGE=${BUILDER_DALI_TF_BASE_MANYLINUX2010}" \
-        .
-    export DALI_TF_BUILDER_CONTAINER_MANYLINUX2010="extract_dali_tf_prebuilt_manylinux2010"
-    nvidia-docker run --name ${DALI_TF_BUILDER_CONTAINER_MANYLINUX2010} ${BUILDER_DALI_TF_MANYLINUX2010} /bin/bash -c 'source /opt/dali/dali_tf_plugin/build_in_custom_op_docker.sh'
-    docker cp "${DALI_TF_BUILDER_CONTAINER_MANYLINUX2010}:/prebuilt/." "prebuilt_manylinux2010"
-    docker rm -f "${DALI_TF_BUILDER_CONTAINER_MANYLINUX2010}"
-
     mkdir -p ./dali_tf_plugin/prebuilt/;
-    cp -r ./prebuilt_manylinux1/* ./prebuilt_manylinux2010/* ./dali_tf_plugin/prebuilt/;
-    rm -rf ./prebuilt_manylinux2010/ ./prebuilt_manylinux1/
+    if [ "${PREBUILD_TF_PLUGINS}" == "YES" ]; then
+        echo "Build image:" ${BUILDER_DALI_TF_MANYLINUX1}
+        docker build -t ${BUILDER_DALI_TF_MANYLINUX1} -f docker/Dockerfile_dali_tf \
+            --build-arg "TF_CUSTOM_OP_BUILDER_IMAGE=${BUILDER_DALI_TF_BASE_MANYLINUX1}" \
+            .
+        export DALI_TF_BUILDER_CONTAINER_MANYLINUX1="extract_dali_tf_prebuilt_manylinux1"
+        nvidia-docker run --name ${DALI_TF_BUILDER_CONTAINER_MANYLINUX1} ${BUILDER_DALI_TF_MANYLINUX1} /bin/bash -c 'source /opt/dali/dali_tf_plugin/build_in_custom_op_docker.sh'
+        docker cp "${DALI_TF_BUILDER_CONTAINER_MANYLINUX1}:/prebuilt/." "prebuilt_manylinux1"
+        docker rm -f "${DALI_TF_BUILDER_CONTAINER_MANYLINUX1}"
+
+        echo "Build image:" ${BUILDER_DALI_TF_MANYLINUX2010}
+        docker build -t ${BUILDER_DALI_TF_MANYLINUX2010} -f docker/Dockerfile_dali_tf \
+            --build-arg "TF_CUSTOM_OP_BUILDER_IMAGE=${BUILDER_DALI_TF_BASE_MANYLINUX2010}" \
+            .
+        export DALI_TF_BUILDER_CONTAINER_MANYLINUX2010="extract_dali_tf_prebuilt_manylinux2010"
+        nvidia-docker run --name ${DALI_TF_BUILDER_CONTAINER_MANYLINUX2010} ${BUILDER_DALI_TF_MANYLINUX2010} /bin/bash -c 'source /opt/dali/dali_tf_plugin/build_in_custom_op_docker.sh'
+        docker cp "${DALI_TF_BUILDER_CONTAINER_MANYLINUX2010}:/prebuilt/." "prebuilt_manylinux2010"
+        docker rm -f "${DALI_TF_BUILDER_CONTAINER_MANYLINUX2010}"
+
+        cp -r ./prebuilt_manylinux1/* ./prebuilt_manylinux2010/* ./dali_tf_plugin/prebuilt/;
+        rm -rf ./prebuilt_manylinux2010/ ./prebuilt_manylinux1/
+    fi
 
     docker build -t ${BUILDER_DALI_TF_SDIST} \
            -f docker/Dockerfile_dali_tf \
-           --build-arg "TF_CUSTOM_OP_BUILDER_IMAGE=${BUILDER_DALI_TF_BASE_MANYLINUX2010}" \
+           --build-arg "TF_CUSTOM_OP_BUILDER_IMAGE=${BUILDER}" \
            --build-arg "NVIDIA_BUILD_ID=${CI_PIPELINE_ID}" \
            --build-arg "NVIDIA_DALI_BUILD_FLAVOR=${DALI_BUILD_FLAVOR}" \
            --build-arg "GIT_SHA=${GIT_SHA}" \
